Lint: add TS rule to enforce replacing Uint8Array with Uint8Array<ArrayBuffer> types
And fix linting support throughout the workspace.

This change is meant to facilitate the future update to TS v5.9,
which will make ArrayBuffers and SharedArrayBuffers incompatible:
https://github.com/microsoft/TypeScript/pull/59417 .

Enforcing Uint8Array<ArrayBuffer> declarations will limit
the need for explicit downcasting of Uint8Arrays
when using them with WebCrypto, Blobs, or other libs
that require ArrayBuffers.